Removed the homesick shell module and folded its code in elsewhere.
